% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{clr}
\alias{clr}
\title{Default scale model discussed in Nixon et al. Beyond Normalizations}
\usage{
clr(X, logWpara, gamma = 0.5)
}
\arguments{
\item{gamma}{variance in scale, documented in Nixon et al. (TODO document
here as well)}
}
\value{
N x nsample matrix
}
\description{
Default scale model discussed in Nixon et al. Beyond Normalizations
}
\author{
Justin Silverman
}
