% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{coef.sm}
\alias{coef.sm}
\title{Scale model where the user may specify the mean and variance for the
log2 fixed effects of the linear (mixed) model.}
\usage{
\method{coef}{sm}(X, logComp, c.mu = NULL, c.cor = NULL)
}
\arguments{
\item{s.mu}{A vector of length P (# covariates; nrow(X))
representing the log2 mean of each sample's scale (cannot be NULL)}

\item{s.cor}{A PxP vector where P is # of covariates (nrow(X))
represeting the covariance matrix in log2 space}
}
\value{
N x nsample matrix
}
\description{
For instance, if you have an intercept and a treatment where the
treatment is expected to increase the log2 scale, then we could
have c.mu=c(0, 1) and c.cor=diag(0.25, 2).
}
\author{
Kyle McGovern
}
