% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldex-summary.R
\name{cohensd}
\alias{cohensd}
\title{Cohen's D}
\usage{
cohensd(m, var)
}
\arguments{
\item{m}{the output of a call to \code{aldex}}

\item{var}{if aldex was called with X being a pre-computed model matrix,
this var should be an integer corresponding to a binary covariate
indicateing the desired effect size to calculate (an effect size between
two groups indicated by the binary covariate). For example, if the third
covariate in the model is an indicator denoting health (0) and disease (1)
then set \code{var=3}. In contrast, if X was a formula (in which case the
\code{data} argument should have been specified) then \code{var} can be set to
the unquoted name of the binary condition variable (e.g.,
\code{var=condition}).}
}
\value{
A (D x nsample)-matrix of Cohen's D statistics for the variable of
interest
}
\description{
Function to compute cohensd on the results provided by the aldex function
}
\details{
WARNING: this function is experimental and requires users read the
documetation fully.
}
\author{
Justin Silverman
}
