% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine.streams}
\alias{combine.streams}
\title{Combine output from aldex streams (internal only)}
\usage{
combine.streams(l)
}
\arguments{
\item{l}{a list, where every element is itself a list of 3D arrays.}
}
\value{
a list of 3D arrays
}
\description{
Takes a list l, where every element is itself a list of 3D arrays. Each
array should have matching first and second dimentions. This function
combines those arrays along the third dimension.
}
\author{
Justin Silverman
}
