% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{default}
\alias{default}
\title{Default scale model discussed in Nixon et al. Beyond Normalizations}
\usage{
default(X, Y, logWpara)
}
\arguments{
\item{gamma}{TODO fix documentation}
}
\value{
N x nsample matrix
}
\description{
Default scale model discussed in Nixon et al. Beyond Normalizations
}
\author{
Justin Silverman
}
