% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{GAMMA.default}
\alias{GAMMA.default}
\title{Default GAMMA Calculation (No External Scale Measurements)}
\usage{
GAMMA.default(logWpara, X, gamma = 0.5)
}
\arguments{
\item{logWpara}{A three-dimensional array (see \code{\link{GAMMA}}).}

\item{X}{A design matrix (see details in \code{\link{GAMMA}}).}

\item{gamma}{Standard deviation(s) for the normal draws; defaults to 0.5.}
}
\value{
A matrix of dimension \code{[ncol(X), nsample]} matching the output of \code{\link{GAMMA}}.
}
\description{
This convenience wrapper calls \code{\link{GAMMA}} assuming you have
a design matrix \code{X} and a single (or vector) \code{gamma},
with \code{externalscale = NULL}.
}
\seealso{
\code{\link{GAMMA}}
}
