% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{GAMMA.sensitivity}
\alias{GAMMA.sensitivity}
\title{Sensitivity Analysis for Multiple Gamma Values (No External Scale)}
\usage{
GAMMA.sensitivity(logWpara, X, gamma_values)
}
\arguments{
\item{logWpara}{A three-dimensional array (see \code{\link{GAMMA}}). This is not user supplied and is generated by calling aldex.lm()}

\item{X}{A design matrix used in \code{\link{default_GAMMA}}.}

\item{gamma_values}{A numeric vector of \code{gamma} values to test.}
}
\value{
A list of matrices; each element is the result of one call to
\code{\link{default_GAMMA}}.
}
\description{
This wrapper calls \code{\link{default_GAMMA}} iteratively over
a set of \code{gamma_values}, returning a list of results for each gamma.
}
\examples{
# Suppose we have logWpara, a design matrix X, and multiple gamma values:
# results <- GAMMA.sensitivity(logWpara, X, c(0.1, 0.5, 1.0))
# str(results)  # Each list element is a matrix for the corresponding gamma
}
\seealso{
\code{\link{default_GAMMA}}, \code{\link{GAMMA}}
}
