% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{GAMMA.sensitivityexternal}
\alias{GAMMA.sensitivityexternal}
\title{Sensitivity Analysis for Multiple Gamma Values (With External Scale)}
\usage{
GAMMA.sensitivityexternal(logWpara, externalscale, gamma_values)
}
\arguments{
\item{logWpara}{A three-dimensional array (see \code{\link{GAMMA}}). This is not user supplied and is generated by calling aldex.lm()}

\item{externalscale}{A numeric vector of means for normal draws (see \code{\link{external_GAMMA}}).}

\item{gamma_values}{A numeric vector of \code{gamma} values to test.}
}
\value{
A list of matrices; each element is the result of one call to
\code{\link{external_GAMMA}}.
}
\description{
This wrapper calls \code{\link{external_GAMMA}} iteratively over
a set of \code{gamma_values}, returning a list of results for each gamma.
}
\details{
Use this when external scale measurements exist but you need to explore
uncertainty (i.e., you do not know the precise measurement SD).
}
\examples{
# Suppose externalscale is c(1, 2, 3), and we want to test multiple gamma values:
# results_ext <- GAMMA.sensitivityexternal(logWpara, c(1,2,3), c(0.1, 0.5, 1.0))
# str(results_ext)  # Each list element is a matrix for the corresponding gamma
}
\seealso{
\code{\link{external_GAMMA}}, \code{\link{GAMMA}}
}
