% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gut_crohns_data}
\alias{gut_crohns_data}
\title{Gut Crohn's microbiome dataset (list)}
\format{
A named list of one matrix and a dataframe:
\describe{
\item{\code{counts}}{A matrix with read counts with 195 rows and 95 columns)}
\item{\code{metadata}}{A data frame with 95 rows and 7 columns containing subject-level covariates.}
}
}
\source{
Vandeputte D, Falony G, Vieira-Silva S, Wang J, Sailer M, Theis S, Raes J (2017).
"Quantitative microbiome profiling links gut community variation to microbial load".
Nature, 551, 507--511. \doi{10.1038/nature24460}
}
\usage{
gut_crohns_data
}
\description{
This dataset contains a matrix and a data frame: genus-level microbiome profiles and
corresponding sample metadata from a Crohn's disease case–control cohort.
The dataset is used in examples and vignettes throughout the package.
}
\details{
The \code{counts} matrix has one row per sample and one column per genus.
The \code{metadata} data frame has one row per sample with metadata, critically Health.status either CD or Control, and Average cell count per gram frozen feces.
}
\keyword{datasets}
