% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{matrix.sm}
\alias{matrix.sm}
\title{This is a function to pass in a custom logScale
matrix. Useful for advanced users defining any
scale model they want or for testing where the
scale is known. Matrix should be on log2 scale.}
\usage{
matrix.sm(X, logComp, custom.logScale)
}
\arguments{
\item{custom.logScale}{A N x nsample matrix of
log2 scale values}
}
\value{
The same N x nsample matrix passed in
if it passes tests, otherwise an error is thrown.
}
\description{
This is a function to pass in a custom logScale
matrix. Useful for advanced users defining any
scale model they want or for testing where the
scale is known. Matrix should be on log2 scale.
}
\author{
Kyle McGovern
}
