% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coda.R
\name{miniclo}
\alias{miniclo}
\title{Closure operation}
\usage{
miniclo(X)
}
\arguments{
\item{X}{should be a D x N matrix, closes along the D dimension}
}
\value{
D x N matrix with columns summing to 1
}
\description{
Closure operation
}
\author{
Justin Silverman
}
