% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oral_mouthwash_data}
\alias{oral_mouthwash_data}
\title{Oral mouthwash example dataset}
\format{
A data frame with \if{html}{\out{<N>}} rows and \if{html}{\out{<K>}} variables:
\describe{
\item{...}{...}
}
}
\source{
<where you got it / how generated>
}
\usage{
oral_mouthwash_data
}
\description{
Example dataset used in vignettes and examples.
}
\keyword{datasets}
