% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{rDirichlet}
\alias{rDirichlet}
\title{Function for sampling Dirichlet random variables}
\usage{
rDirichlet(n, alpha)
}
\arguments{
\item{n}{number of samples}

\item{alpha}{D-vector of Dirichlet parameters}
}
\value{
an D x n matrix of samples
}
\description{
Function for sampling Dirichlet random variables
}
\author{
Justin Silverman
}
