% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{rDirichletMat}
\alias{rDirichletMat}
\title{Function for sampling Dirichlet random variables}
\usage{
rDirichletMat(n, Alpha)
}
\arguments{
\item{n}{number of samples}

\item{Alpha}{DxN matrix of Dirichlet parameters}
}
\value{
an D x n matrix of samples
}
\description{
Function for sampling Dirichlet random variables
}
\author{
Justin Silverman
}
