% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{req}
\alias{req}
\title{Test object contains elements}
\usage{
req(obj, names)
}
\arguments{
\item{obj}{object (list type)}

\item{names}{names of elements required to be in the list}
}
\description{
Test object contains elements
}
\author{
Justin Silverman
}
