% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirichlet.R
\name{rLogDirichlet}
\alias{rLogDirichlet}
\title{Function for sampling Dirichlet random variables (base-2 normalized via log-sum-exp for stability)}
\usage{
rLogDirichlet(n, alpha)
}
\arguments{
\item{n}{number of samples}

\item{alpha}{D-vector of Dirichlet parameters}
}
\value{
a D x n matrix of samples
}
\description{
Function for sampling Dirichlet random variables (base-2 normalized via log-sum-exp for stability)
}
\author{
Justin Silverman
}
