% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{aldex.lm.sim.clr}
\alias{aldex.lm.sim.clr}
\title{Simulation function soley for testing and exploring ALDEx2, Truth is in CLR
Coordinates.}
\usage{
aldex.lm.sim.clr(D = 10, N = 11, P = 2, depth = 10000)
}
\arguments{
\item{D}{number of taxa/genes}

\item{N}{number of samples}

\item{P}{number of covariates}

\item{depth}{sum of counts for each multinomial draw}
}
\value{
a list with elements Y, X, W, and Lambda
}
\description{
Not designed to create realistic data. Does not add any noise to linear
regression! True W is in CLR Corrdinates
}
\author{
Justin Silverman
}
