% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aldex.R
\name{aldex}
\alias{aldex}
\title{ALDEx3 Linear Modles}
\usage{
aldex(
  Y,
  X,
  data = NULL,
  nsample = 2000,
  GAMMA = NULL,
  streamsize = 8000,
  return.pars = c("X", "estimate", "std.error", "p.val", "p.val.adj"),
  return.samples = FALSE,
  p.adjust.method = "BH"
)
}
\arguments{
\item{X}{either a formula (in which case DATA must be non-null) or a model
matrix of dimension P x N (P is number of linear model covariates). If a
formula is passed it should not include the target Y, e.g., should simply
be "~condition-1" (note the lack of the left hand side).}

\item{data}{a data frame for use with formula, must have N rows}

\item{nsample}{number of monte carlo replicates}

\item{GAMMA}{the scale model, can be a function or an N x nsample matrix. If
a function is passed, it should take one argument (pi) which is a N x D x
nsample array of log-transformed relative abundances. That function must
in turn output an N x nsample matrix of scale samples on the log2 scale.}

\item{streamsize}{(default 8000) memory footprint (approximate) at which to
use streaming. This should be thought of as the number of Mb for each
streaming chunk. If D\emph{N}nsample*8/1000000 is less than streamsize then no
streaming will be performed. Note, to conserve memory, samples from the
Dirichlet and scale models will not be returned if streaming is used.
Streaming can be turned off by setting streamsize=Inf.}

\item{return.pars}{what results should be returned, see return section below.

.samples (default TRUE) if true, return samples for logWpara composition and
logWperp (scale). Will override to FALSE if streaming is required.}

\item{p.adjust.method}{(default BH) The method for multiple hypothesis test
correction. See \code{p.adjust} for all available methods.}

\item{counts}{an (D x N) matrix of sequence count, N is number of samples, D
is number of taxa or genes}
}
\value{
a list with elements controled by parameter resturn.pars. Options
include:
\itemize{
\item X: P x N covariate matrix
\item estimate: (P x D x nsample array) of linear model estimates
\item std.error: (P x D x nsample array) of standard error for the estimates
\item p.val: (P x D) matrix, unadjusted p-value for two-sided t-test
\item p.val.adj: (P x D) matrix, p-value for two-sided t-test adjusted
according to \code{p.adj.method}
\item logWperp: (N x S) matrix, samples of the log scale from the scale model
\item logWpara: (D x N x S) array, samples of the log composition from
the multinomial-Dirichlet
\item streaming: boolean, detnote if streaming was used.
Note, logWperp and logWpara are not returned if streaming is active
}
}
\description{
ALDEx3 Linear Models
}
\author{
Justin Silverman
}
