% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fflm.R
\name{fflm}
\alias{fflm}
\title{Freaking Fask Linear Models}
\usage{
fflm(Y, X, hce = FALSE)
}
\arguments{
\item{Y}{a numeric array (N x D X S) where D is the number of taxa/genes, N
is the number of samples, and S is the number of posterior samples}

\item{X}{a numeric matrix (N x P) where P is number of covariates}

\item{hce}{(default FALSE), should Heteroskedasticity-consistent standard
errors be used? Slower to compute.}
}
\value{
A list of (P x D x S)-arrays with the OLS point estimates, the
standard errors, and the two-sided p-values for each coefficient (P), of
each model fit to each taxa (D) and each posterior sample (S)
}
\description{
Tailored for ALDEx3 where covariates are shared between massive numbers of
linear regressions where only Y is changing.
}
\author{
Justin Silverman
}
