% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{sample.sm}
\alias{sample.sm}
\title{Scale model where the user may specify the mean and variance of the
log2 of the scale for each individual sample.}
\usage{
sample.sm(X, logComp, s.mu = NULL, s.var = NULL, s.cor = NULL)
}
\arguments{
\item{s.mu}{A vector of length N (# samples; ncol(X))
representing the log2 mean of each sample's scale (cannot be NULL)}

\item{s.var}{A vector of length N (# samples; ncol(X))
representing the variance of each sample in log2 space (cannot be
NULL if s.cor is NULL)}

\item{s.cor}{A NxN vector where N is # of samples (ncol(X)) representing
the covariance matrix in log2 space (cannot be NULL if s.var is NULL)}
}
\value{
N x nsample matrix
}
\description{
Use this scale model if you want to specify the mean and (co)variance
of the log2 scale. This scale model is particularly useful if you have
paired qPCR or flow cytometry measurements of scale. Broadly useful if
you want to specify the scale for each individual sample.
}
\author{
Kyle McGovern
}
